add download

+++ updater.py
@@ -2,6 +2,7 @@ import os
 import json
 
 from minio import Minio
+from minio.error import ResponseError
 from compare import make_hashs, has_changed, PATH, is_new
 
 
@@ -14,6 +15,19 @@ class CookieUpdater:
                             access_key=self.credentials['key_id'],
                             secret_key=self.credentials['key'],
                             region=self.credentials['region'])
+
+    def update_local(self, folder):
+        objects = self.client.list_objects(self.credentials['bucket'], recursive=True)
+        for obj in objects:
+            print(obj.object_name)
+
+            try:
+                data = self.client.get_object(self.credentials['bucket'], obj.object_name)
+                with open('{}/{}'.format(folder, obj.object_name), 'wb') as file_data:
+                    for d in data.stream(32 * 1024):
+                        file_data.write(d)
+            except ResponseError as err:
+                print(err)
 
     def upload_file(self, filename):
         print('file to upload: {}'.format(filename))
@@ -34,36 +48,36 @@ class CookieUpdater:
         except Exception as e:
             print('error uploading {}:\n{}'.format(filename, e))
 
-    def check_new_cookies(self):
-        pass
-
     def run(self):
-        changed_files = list()
-        new_cookies = list()
+        if os.path.isfile('{}{}'.format(PATH, 'INIT')):
+            self.update_local(PATH)
+        else:
+            changed_files = list()
+            new_cookies = list()
 
-        for path, subdirs, files in os.walk(PATH):
-            print(path, subdirs, files)
-            for filename in files:
-                filename = '{}{}'.format(PATH, filename)
-                if '.sha256' not in filename:
-                    if is_new(filename):
-                        new_cookies.append(filename)
-                    elif has_changed(filename):
-                        print('FILE: {}'.format(filename))
-                        changed_files.append(os.path.join(filename))
-                    else:
-                        print('not new neither has changed')
+            for path, subdirs, files in os.walk(PATH):
+                print(path, subdirs, files)
+                for filename in files:
+                    filename = '{}{}'.format(PATH, filename)
+                    if '.sha256' not in filename:
+                        if is_new(filename):
+                            new_cookies.append(filename)
+                        elif has_changed(filename):
+                            print('FILE: {}'.format(filename))
+                            changed_files.append(os.path.join(filename))
+                        else:
+                            print('not new neither has changed')
 
-        print('changed: {}'.format(changed_files))
-        print('new: {}'.format(new_cookies))
+            print('changed: {}'.format(changed_files))
+            print('new: {}'.format(new_cookies))
 
-        for new_cookie in new_cookies:
-            self.upload_file(new_cookie)
+            for new_cookie in new_cookies:
+                self.upload_file(new_cookie)
 
-        for cookie in changed_files:
-            self.upload_file(cookie)
-        make_hashs(PATH)
-        print('\n' + '*' * 20 + '\n')
+            for cookie in changed_files:
+                self.upload_file(cookie)
+            make_hashs(PATH)
+            print('\n' + '*' * 20 + '\n')
